add 4/1

+++ ShopAnDam/ShopAnDam/Models/Framework/Order_Detail.cs
@@ -10,7 +10,7 @@ namespace ShopAnDam.Models.Framework
     {
         public int ID { get; set; }
 
-        public int OrderID { get; set; }
+        public long OrderID { get; set; }
 
         public int ProductID { get; set; }
